Rename TreeNode.js to _TreeNode.js as it is not an exported component
